IT IS WORKING!

+++ policy_based/goexplore_py/data_classes.py
@@ -24,7 +24,7 @@ except ModuleNotFoundError:
 @dataclass
 class GridDimension:
     attr: str
-    div: int
+    div: float
 
 
 @dataclass
